pyproject.toml
[project]
name = "codebase-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.17",
    "langchain>=0.3.18",
    "neo4j>=5.28.1",
    "networkx>=3.4.2",
    "openai>=1.63.0",
    "py2neo>=2021.2.4",
    "tiktoken>=0.9.0",
    "jedi>=0.19.2",
    "libcst>=1.6.0",
    "pydot>=3.0.4",
    "langchain-neo4j>=0.3.0",
    "langchain-openai>=0.3.6",
    "streamlit>=1.42.2",
    "rarfile>=4.2",
    "boto3>=1.37.9",
    "fastapi>=0.115.11",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "skforecast>=0.15.0",
]
